chore: fix monorepo imports
