+++ packages/client-core/package.json
@@ -13,6 +13,7 @@
   },
   "license": "MIT",
   "main": "./dist/index.cjs",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "build-bundle": "rollup --config rollup.config.ts --configPlugin typescript",
     "build-types-check": "tsc --project tsconfig.check.json",
