+++ packages/client-core/src/__tests__/index.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-describe('Fetch API client', () => {
+describe('client core', () => {
   it('works', () => {
     expect(1).toBe(1);
   });
